requirements.txt
Django==1.6.3
Markdown==2.4
django-filter==0.7
djangorestframework==2.3.13
